refactor: move constants folder to data
