+++ src/components/Toast/toast.tsx
@@ -13,22 +13,14 @@ type ToastPropsType = {
 const Toast = ({ message }: ToastPropsType) => {
   const [open, setOpen] = React.useState(true);
 
-  const handleClose = (event?: React.SyntheticEvent | Event, reason?: string) => {
-    if (reason === 'clickaway') {
-      return;
-    }
-
-    setOpen(false);
-  };
-
   return (
     <Snackbar
       open={open}
       autoHideDuration={4000}
-      onClose={handleClose}
+      onClose={() => setOpen(false)}
       anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
     >
-      <Alert onClose={handleClose} severity="error" sx={{ width: '100%' }}>
+      <Alert onClose={() => setOpen(false)} severity="error" sx={{ width: '100%' }}>
         {message}
       </Alert>
     </Snackbar>
