+++ src/data/constants/validation.ts
@@ -1,0 +1,9 @@
+export const REGEXP_PASSWORD_VALID_CHARACTERS = /^[\w!@+#%({}\[\]\\/)=$^<>&*-]{8,}$/;
+
+export const REGEXP_SPECIAL_CHARACTERS = /[!@_+#%({}\[\]\\/)=$^<>&*-]/;
+
+export const MIN_PASSWORD_LENGTH = 8;
+
+export const REGEXP_LOGIN_VALID_CHARACTERS = /^[\w]{3,}$/;
+
+export const REGEXP_NAME_VALID_CHARACTERS = /^[А-ЯЁа-яёA-Za-z\s-]{2,}$/;
